Split config.toml into node.toml and controller.toml
