--- mipengine/config.toml
@@ -1,19 +1,0 @@
-[logging]
-log_level = "INFO"
-
-[rabbitmq]
-user = "user"
-password = "password"
-vhost = "user_vhost"
-
-[monetdb]
-database = "db"
-username = "monetdb"
-password = "monetdb"
-
-[node]
-identifier = "localnode1"
-tasks_soft_time_limit = 30
-tasks_time_limit = 60
-run_udf_soft_time_limit = 60
-run_udf_time_limit = 120
