+++ mipengine/controller.toml
@@ -1,0 +1,1 @@
+log_level = "INFO"
