+++ mipengine/controller/__init__.py
@@ -1,0 +1,8 @@
+import toml
+from importlib.resources import open_text
+
+import mipengine
+from mipengine.utils import AttrDict
+
+with open_text(mipengine, "controller.toml") as fp:
+    config = AttrDict(toml.load(fp))
