+++ mipengine/node.toml
@@ -1,0 +1,20 @@
+identifier = "localnode1"
+
+log_level = "INFO"
+
+[celery]
+worker_concurrency = 1
+task_soft_time_limit = 30
+task_time_limit = 60
+run_udf_soft_time_limit = 60
+run_udf_time_limit = 120
+
+[rabbitmq]
+user = "user"
+password = "password"
+vhost = "user_vhost"
+
+[monetdb]
+database = "db"
+username = "monetdb"
+password = "monetdb"
