--- mipengine/node/celery_config.py
@@ -1,5 +1,0 @@
-from mipengine import config
-
-worker_concurrency = 1
-task_soft_time_limit = config.node.tasks_soft_time_limit
-task_time_limit = config.node.tasks_time_limit
