+++ mipengine/node/node.py
@@ -2,21 +2,20 @@ import argparse
 
 from celery import Celery
 
-from mipengine import config
+from mipengine.node import config
 from mipengine.common.node_catalog import node_catalog
-from mipengine.node import celery_config
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--node-id", help="Current node identifier.", required=True)
     extra_args, celery_args = parser.parse_known_args()
-    config.node.identifier = extra_args.node_id
+    config.identifier = extra_args.node_id
 
 global_node = node_catalog.get_global_node()
-if global_node.nodeId == config.node.identifier:
+if global_node.nodeId == config.identifier:
     current_node = global_node
 else:
-    current_node = node_catalog.get_local_node(config.node.identifier)
+    current_node = node_catalog.get_local_node(config.identifier)
 
 rabbitmqURL = current_node.rabbitmqURL
 user = config.rabbitmq.user
@@ -37,7 +36,9 @@ app = Celery(
     ],
 )
 
-app.config_from_object(celery_config)
+app.conf.worker_concurrency = config.celery.worker_concurrency
+app.conf.task_soft_time_limit = config.celery.task_soft_time_limit
+app.conf.task_time_limit = config.celery.task_time_limit
 
 if __name__ == "__main__":
     app.worker_main(celery_args)
