+++ mipengine/node/tasks/tables.py
@@ -4,8 +4,8 @@ from celery import shared_task
 
 from mipengine.common.node_tasks_DTOs import TableInfo
 from mipengine.common.node_tasks_DTOs import TableSchema
+from mipengine.node import config
 from mipengine.node.monetdb_interface import tables
-from mipengine.node.monetdb_interface.common_actions import config
 from mipengine.node.monetdb_interface.common_actions import create_table_name
 
 
@@ -43,9 +43,7 @@ def create_table(context_id: str, command_id: str, schema_json: str) -> str:
         The name of the created table in lower case
     """
     schema_object = TableSchema.from_json(schema_json)
-    table_name = create_table_name(
-        "table", command_id, context_id, config["node"]["identifier"]
-    )
+    table_name = create_table_name("table", command_id, context_id, config.identifier)
     table_info = TableInfo(table_name.lower(), schema_object)
     tables.create_table(table_info)
     return table_name.lower()
