+++ mipengine/node/tasks/views.py
@@ -3,7 +3,7 @@ from typing import List
 from celery import shared_task
 
 from mipengine.node.monetdb_interface import views
-from mipengine.node.monetdb_interface.common_actions import config
+from mipengine.node import config
 from mipengine.node.monetdb_interface.common_actions import create_table_name
 
 
@@ -54,9 +54,7 @@ def create_view(
     str
         The name of the created view in lower case
     """
-    view_name = create_table_name(
-        "view", command_id, context_id, config["node"]["identifier"]
-    )
+    view_name = create_table_name("view", command_id, context_id, config.identifier)
     views.create_view(
         view_name=view_name, pathology=pathology, datasets=datasets, columns=columns
     )
