+++ tests/integration_tests/nodes_communication.py
@@ -1,7 +1,7 @@
 from celery import Celery
 
 from mipengine.common.node_catalog import node_catalog
-from mipengine import config
+from mipengine.node import config
 
 
 def get_celery_app(node_id: str):
